Some fixes.

+++ include/extern.h
@@ -3569,7 +3569,7 @@ E int FDECL(wand_skill_hit_bonus, (int));
 E void FDECL(skill_init, (const struct def_skill*, const struct def_skill *));
 E int FDECL(martial_arts_multishot_percentage_chance, (int));
 E int FDECL(get_skill_critical_strike_chance, (enum p_skills, boolean));
-E int FDECL(get_wand_damage_multiplier, (boolean));
+E double FDECL(get_wand_damage_multiplier, (boolean));
 E int FDECL(get_exceptionality_multiplier, (int));
 
 /* ### were.c ### */
